requirements-adk.txt
# Google ADK + Vertex (Agent Engine deploy / local adk run). Install into project venv:
#   python -m venv .venv
#   .\.venv\Scripts\pip install -r requirements-adk.txt
google-adk[a2a]>=1.27.0
google-cloud-aiplatform>=1.70.0
google-cloud-bigquery>=3.0.0
# Local dev + cost_metrics_agent tools (Agent Engine also merges from agent requirements.txt)
psycopg[binary]>=3.2
